+++ fixtures/classes/baseclass-component.cs
@@ -1,3 +1,4 @@
+// This is a header
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Runtime.Serialization;