+++ source/compile.ts
@@ -6,6 +6,7 @@ import { CompilerOptions } from "./compiler-types";
 
 const defaultOptions = {
   generator: generateCsharp,
+  header: "",
   indent: 2,
   namespace: "",
   parser: parsePropTypes,
@@ -14,7 +15,7 @@ const defaultOptions = {
 
 export default function compile(sourceCode: string, options?: CompilerOptions) {
   const opts = { ...defaultOptions, ...(options || {}) };
-  const { generator, indent, namespace, parser, supertype } = opts;
+  const { generator, header, indent, namespace, parser, supertype } = opts;
   assert(typeof parser === "function", "Options.parser is not a function.");
   assert(
     typeof generator === "function",
@@ -27,6 +28,7 @@ export default function compile(sourceCode: string, options?: CompilerOptions) {
   const { typeName, types, supertype: parsedSupertype } = parseResult;
 
   const code = generator(types, typeName, {
+    header,
     indent,
     namespace,
     supertype: parsedSupertype || supertype