+++ source/generators/typescript/index.ts
@@ -7,7 +7,7 @@ import { GeneratorOptions } from "../../compiler-types";
 export default function typescript(
   types: TypeTree,
   typeName: string,
-  { supertype, indent = 2, namespace }: GeneratorOptions = {}
+  { header, supertype, indent = 2, namespace }: GeneratorOptions = {}
 ) {
   const typeString = generateInterface(typeName, types, supertype);
 
@@ -18,8 +18,9 @@ export default function typescript(
   const importsString = [...(supertype ? [supertype] : []), ...componentImports]
     .map(i => `import { ${i} } from "./${i}";`)
     .join("\n");
+  const headerString = header ? `${header}\n` : "";
 
-  const fileContent = `${importsString}\n\n${typesWithNamespace}\n`;
+  const fileContent = `${headerString}${importsString}\n\n${typesWithNamespace}\n`;
 
   return indentBraces(fileContent, indent);
 }